separación de la clase libro
